lsif-clang: Reduce contention + skip repeat work for files. (#85)
* lsif-clang: Reduce contention by using per-table mutexes.

For an indexing job of CMake on a 60 core machine:
- Before this change: indexing time 30s with about 31/60 core utilization
- After this change:  indexing time 25s with about 42/60 core utilization

There doesn't seem to be much effect on 16 core machines.

This was changed in upstream LLVM too (in clangd).
The patch to IndexerMain.cpp is not used, because the code is
duplicated in LSIFClangMain.cpp, but we keep the two in sync.

* lsif-clang: Add file skipping optimization from upstream patch.

This cuts down build time further by about 40% for CMake.

Practically a cherry-pick of 9cb00e7133ea3da6a49ade95e3708240c2aaae39